+++ CHANGES.md
@@ -16,6 +16,7 @@
 - Add OpenXRFbPassthroughGeometry node
 - Add OpenXRMetaPassthroughColorLut
 - Add feature flags to Khronos loader for HTC
+- Add XR_HTC_passthrough extension wrapper
 
 ## 2.0.3
 - Migrate the export scripts from gdscript to C++ via gdextension